==> 1. Cleaning.ipynb <==
# **Cleaning**

import pandas as pd
import warnings
warnings.filterwarnings('ignore')

from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
file_path = '/content/drive/MyDrive/Colab/Ulasan_Coretax.csv'
df = pd.read_csv(file_path)
df.head()

# Cek info data
df.info()

# Cek nilai duplikasi dalam data
df.duplicated().sum()

# 1. Hapus baris dengan username null
df = df.dropna(subset=['username'])

# 2. Hapus baris dengan username 'kring_pajak'
df = df[df['username'].str.lower() != 'kring_pajak']

# 3. Hapus duplikasi
df = df.drop_duplicates()

# 4. Reset index agar rapi
df = df.reset_index(drop=True)

# 5. Ubah kolom created_at ke format datetime yang benar
df['created_at'] = pd.to_datetime(df['created_at'], errors='coerce')

# 6. Drop baris jika ada created_at yang gagal diubah (jadi NaT)
df = df.dropna(subset=['created_at'])

# 7. Ubah format tanggal menjadi 'dd Mon yyyy'
df['created_at'] = df['created_at'].dt.strftime('%d %b %Y')

# 8. Rename kolom agar konsisten
df.rename(columns={
    'created_at': 'tanggal',
    'full_text': 'ulasan',
    'username': 'username'
}, inplace=True)
df.head()

import pandas as pd
import re
import string

def clean_text(text):
  if isinstance(text, str):

    # Mengubah teks menjadi huruf kecil
    text = text.lower()

    #Mengganti simbol menjadi spasi
    text = text.replace('.','').replace('/','').replace(',','').replace('-','')

    # Menghapus mention
    text = re.sub(r'@\w+', '', text)

    # Menghapus hashtag
    text = re.sub(r'#\w+', '', text)

    # Menghapus URL/link
    text = re.sub(r'http\S+|www\S+|https\S+', '', text)

    # Menghapus angka
    text = re.sub(r'\d+', '', text)

    # Menghapus emoji dan karakter non-ASCII
    text = re.sub(r'[^\x00-\x7F]+', '', text)

    # Menghapus karakter non-alphanumeric
    text = re.sub(r'[^a-zA-Z0-9\s]', '', text)

    # Menghapus tanda baca
    text = text.translate(str.maketrans('', '', string.punctuation))

    # Menghapus spasi berlebih
    text = re.sub(r'\s+', ' ', text).strip()
    return text
  else:
    return ''

# Terapkan ke kolom ulasan
df['ulasan'] = df['ulasan'].fillna('')
df ['cleaning'] = df ['ulasan'].apply(clean_text)
df.head()

# Cek info data
df.info()

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(1)Hasil_Cleaning.csv', index=False)
