==> 3. Stopword Removal.ipynb <==
# **Stopword Removal**

import pandas as pd

from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
df = pd.read_csv('/content/drive/MyDrive/Colab/(2)Hasil_Normalization.csv', keep_default_na=False)
df.head()

# Membaca kamus stopword dari file .csv yang telah di upload di Google Drive
stopwords_path = "/content/drive/MyDrive/Colab/kamus_stopword.csv"
df_stopwords = pd.read_csv(stopwords_path, header=None)
stopwords_indonesia = df_stopwords[0].tolist()

def remove_stopwords(text):
    if isinstance(text, str):
        words = text.split()
        filtered_words = []
        for word in words:
            if word.lower() not in stopwords_indonesia:
                filtered_words.append(word)
        filtered_text = ' '.join(filtered_words)
        return filtered_text
    else:
        return text

# Menghapus stopword pada kolom 'normalization' dan menyimpan hasilnya dalam kolom 'stopword'
df['stopword'] = df['normalization'].apply(remove_stopwords)
df.head()

# Cek info data
df.info()

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(3)Hasil_Stopword.csv', index=False)
