==> 4. Stemming.ipynb <==
# **Stemming**

!pip install Sastrawi
import pandas as pd
import nltk
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
df = pd.read_csv('/content/drive/MyDrive/Colab/(3)Hasil_Stopword.csv', keep_default_na=False)
df.head()

# Inisialisasi stemmer
stemmer = StemmerFactory().create_stemmer()

# Fungsi untuk melakukan lemmatisasi pada teks
def lemmatize_text(text):
    return stemmer.stem(text)

# Melakukan stemming pada kolom 'stopword'
df['stemming'] = df['stopword'].astype(str).apply(lambda x: stemmer.stem(x))
df.head()

# Cek info data
df.info()

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(4)Hasil_Stemming.csv', index=False)
