==> 5. Tokenizing.ipynb <==
# **Tokenizing**

import pandas as pd

from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
df = pd.read_csv('/content/drive/MyDrive/Colab/(4)Hasil_Stemming.csv', keep_default_na=False)
df.head()

import nltk
nltk.download('punkt')

from nltk.tokenize import word_tokenize

# Fungsi untuk melakukan tokenisasi
def tokenize_text(text):
    if isinstance(text, str):
        return word_tokenize(text)
    else:
        return []

# Terapkan tokenisasi pada kolom 'stemming'
df['tokenizing'] = df['stemming'].apply(lambda x: x.split() if isinstance(x, str) else [])
df.head()

# Cek info data
df.info()

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(5)Hasil_Tokenizing.csv', index=False)
