==> 6. Translation.ipynb <==
# **Translation**

import pandas as pd
from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
df = pd.read_csv('/content/drive/MyDrive/Colab/(5)Hasil_Tokenizing.csv', keep_default_na=False)
df.head()

!pip install deep-translator

from deep_translator import GoogleTranslator

# Fungsi untuk translate
def translate_to_english(text):
    # Jika string kosong, langsung kembalikan tanpa translate
    if not isinstance(text, str) or text.strip() == '':
        return ''  # Atau bisa juga: return 'no text'
    try:
        translated = GoogleTranslator(source='auto', target='en').translate(text)
        return translated if translated else ''
    except:
        return ''  # Jika error translasi, tetap return string kosong

# Menerapkan pada kolom tokenizing
df['translation'] = df['tokenizing'].astype(str).apply(translate_to_english)
df.head()

# Cek info data
df.info()

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(6)Hasil_Translation.csv', index=False)
