==> 7. Labelling.ipynb <==
# **Labelling**

import pandas as pd
from google.colab import drive
drive.mount('/content/drive')

# Memanggil data yang berada di Google Drive
df = pd.read_csv('/content/drive/MyDrive/Colab/(6)Hasil_Translation.csv', keep_default_na=False)
df.head()

!pip install tweet-preprocessor
!pip install textblob
!pip install wordcloud
!pip install nltk

import preprocessor as p
from textblob import TextBlob
import nltk
from nltk.tokenize import word_tokenize
from nltk.stem import PorterStemmer

nltk.download('punkt')

data_tweet = list(df['translation'])

polaritas = 0
status = []
total_positif = total_negatif = total = 0

for i, tweet in enumerate(data_tweet):
    if isinstance(tweet, str):
        analysis = TextBlob(tweet)
        polaritas += analysis.sentiment.polarity

        if analysis.sentiment.polarity > 0.0:
            total_positif += 1
            status.append('positif')
        else:
            total_negatif += 1
            status.append('negatif')

        total += 1

df['sentimen'] = status
df.head()

# Cek info data
df.info()

sentiment_counts = df['sentimen'].value_counts()
print("Jumlah Sentimen:")
print(sentiment_counts)

from google.colab import drive
drive.mount('/content/drive')

# Simpan ke folder di Google Drive
df.to_csv('/content/drive/MyDrive/Colab/(7)Hasil_Labelling.csv', index=False)
